requirements.txt
langchain==0.0.123
openai==0.27.0
python-dotenv==1.0.0
streamlit==1.19.0
langgraph==0.1.2
